oflline/requirements.txt
tensorflow==2.13.0
pycryptodome
opencv-python
pillow
espeak